updated e2e testcase
