+++ mobile-app/cypress/e2e/functional/wallet/settings/networkDetails.spec.ts
@@ -3,10 +3,10 @@ import dayjs from 'dayjs'
 context('Wallet - Network detail screen - outside wallet context', () => {
   beforeEach(() => {
     cy.visit('/')
-    cy.blockAllFeatureFlag()
     cy.exitWallet()
-    cy.getByTestID('create_wallet_button').click()
-    cy.getByTestID('guidelines_switch').click()
+    cy.getByTestID('get_started_button').click()
+    cy.getByTestID('guidelines_check').click()
+    cy.url().should('include', 'wallet/onboarding/guidelines')
     cy.restoreLocalStorage()
   })
 
@@ -15,44 +15,35 @@ context('Wallet - Network detail screen - outside wallet context', () => {
   })
 
   it('should check network detail without switching network', function () {
-    cy.getByTestID('create_recovery_words_button').click()
-    cy.url().should('include', 'wallet/mnemonic/create')
-    cy.getByTestID('header_active_network').first().invoke('text').then((network) => {
-      cy.getByTestID('header_status_indicator').should('have.css', 'background-color').then((statusBgColor) => {
-        cy.getByTestID('wallet_header_container').filter(':visible').click()
-        cy.getByTestID('network_details_network').should('exist').contains(network)
-        cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', statusBgColor)
+    cy.getByTestID('header_active_network').first().invoke('text').then((network: string) => {
+      cy.getByTestID('header_network_icon').find('path').should('have.css', 'fill').then((statusBgColor) => {
+        cy.getByTestID('header_active_network').first().filter(':visible').click()
+        cy.getByTestID(`button_network_${network}_check`).should('exist')
+        cy.getByTestID(`button_network_${network}_check`).should('have.css', 'color', statusBgColor)
       })
     })
   })
 
   it('should check network detail by switching network', function () {
-    cy.getByTestID('create_recovery_words_button').click()
-    cy.url().should('include', 'wallet/mnemonic/create')
-    cy.getByTestID('header_active_network').first().invoke('text').then((network) => {
-      cy.getByTestID('header_status_indicator').should('have.css', 'background-color').then((statusBgColor) => {
-        cy.getByTestID('wallet_header_container').filter(':visible').click()
-        cy.getByTestID('network_details_network').should('exist').contains(network)
-        cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', statusBgColor)
-        cy.getByTestID('network_details_header_back').click()
-        cy.url().should('include', 'wallet/mnemonic/create')
+    cy.getByTestID('header_active_network').first().invoke('text').then((network: string) => {
+      cy.getByTestID('header_network_icon').find('path').should('have.css', 'fill').then((statusBgColor) => {
+        cy.getByTestID('header_active_network').first().filter(':visible').click()
+        cy.getByTestID(`button_network_${network}_check`).should('exist')
+        cy.getByTestID(`button_network_${network}_check`).should('have.css', 'color', statusBgColor)
         cy.go('back')
         cy.url().should('include', 'wallet/onboarding/guidelines')
         cy.getByTestID('header_active_network').first().click()
-        cy.getByTestID('onboarding_network_selection_screen').should('exist')
+        cy.getByTestID('network_details').should('exist')
         cy.getByTestID('button_network_Playground').click()
         cy.go('back').wait(3000)
         cy.url().should('include', 'wallet/onboarding/guidelines')
-        cy.getByTestID('guidelines_switch').click()
-        cy.getByTestID('create_recovery_words_button').click()
-        cy.url().should('include', 'wallet/mnemonic/create')
-        cy.getByTestID('header_active_network').first().invoke('text').then((updatedNetwork) => {
-          cy.getByTestID('header_status_indicator').should('have.css', 'background-color').then((updatedStatusBgColor) => {
-            cy.getByTestID('wallet_header_container').filter(':visible').click()
+        cy.getByTestID('header_active_network').first().invoke('text').then((updatedNetwork: string) => {
+          cy.getByTestID('header_network_icon').find('path').should('have.css', 'fill').then((updatedStatusBgColor) => {
             expect(network).not.eq(updatedNetwork)
-            cy.getByTestID('network_details_status_value').should('exist').contains('Connected')
-            cy.getByTestID('network_details_network').should('exist').contains(updatedNetwork)
-            cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', updatedStatusBgColor)
+            cy.getByTestID('header_active_network').first().filter(':visible').click()
+            cy.getByTestID(`button_network_${network}_uncheck`).should('exist')
+            cy.getByTestID(`button_network_${updatedNetwork}_check`).should('exist')
+            cy.getByTestID(`button_network_${updatedNetwork}_check`).should('have.css', 'color', updatedStatusBgColor)
           })
         })
       })
@@ -60,8 +51,6 @@ context('Wallet - Network detail screen - outside wallet context', () => {
   })
 
   it('should check network details after failed API calls for stats', function () {
-    cy.getByTestID('create_recovery_words_button').click()
-    cy.url().should('include', 'wallet/mnemonic/create')
     cy.intercept('**/regtest/stats', {
       statusCode: 404,
       body: '404 Not Found!',
@@ -70,22 +59,17 @@ context('Wallet - Network detail screen - outside wallet context', () => {
       }
     })
     cy.wait(3000)
-    cy.getByTestID('header_active_network').first().invoke('text').then((network) => {
-      cy.getByTestID('header_status_indicator').should('have.css', 'background-color').then((statusBgColor) => {
-        cy.getByTestID('wallet_header_container').filter(':visible').click()
-        cy.getByTestID('network_details_network').should('exist').contains(network)
-        cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', statusBgColor)
-        cy.getByTestID('network_details_status_value').should('exist').contains('Disconnected')
-        cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', 'rgb(239, 68, 68)')
-        cy.getByTestID('network_details_block_height').should('exist').contains(0)
-        cy.getByTestID('network_details_total_masternodes').should('exist').contains(0)
-      })
+    cy.getByTestID('header_active_network').first().invoke('text').then((network: string) => {
+      cy.getByTestID('header_network_icon').find('path').should('have.css', 'fill', 'rgb(229, 69, 69)')
+      cy.getByTestID('header_active_network').first().filter(':visible').click()
+      cy.getByTestID('network_details_network').should('exist').contains(network)
+      cy.getByTestID(`button_network_${network}_check`).should('exist')
+      cy.getByTestID('network_details_block_height').should('exist').contains(0)
+      cy.getByTestID('network_details_total_masternodes').should('exist').contains(0)
     })
   })
 
   it('should check network details after intercepting stats call', function () {
-    cy.getByTestID('create_recovery_words_button').click()
-    cy.url().should('include', 'wallet/mnemonic/create')
     cy.intercept('**/regtest/stats', {
       statusCode: 200,
       body: {
@@ -103,11 +87,11 @@ context('Wallet - Network detail screen - outside wallet context', () => {
       }
     })
     cy.wait(3000)
-    cy.getByTestID('header_active_network').first().invoke('text').then((network) => {
-      cy.getByTestID('header_status_indicator').should('have.css', 'background-color').then((statusBgColor) => {
-        cy.getByTestID('wallet_header_container').filter(':visible').click()
+    cy.getByTestID('header_active_network').first().invoke('text').then((network: string) => {
+      cy.getByTestID('header_network_icon').find('path').should('have.css', 'fill').then((statusBgColor) => {
+        cy.getByTestID('header_active_network').first().filter(':visible').click()
         cy.getByTestID('network_details_network').should('exist').contains(network)
-        cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', statusBgColor)
+        cy.getByTestID(`button_network_${network}_check`).should('have.css', 'color', statusBgColor)
         cy.getByTestID('network_details_block_height').should('exist').contains(100)
         cy.getByTestID('network_details_total_masternodes').should('exist').contains(10)
         cy.getByTestID('network_details_last_sync').invoke('text').then((lastSuccessfulSync) => {
@@ -124,46 +108,42 @@ context('Wallet - Network detail screen - outside wallet context', () => {
 context('Wallet - Network detail screen - with wallet context', () => {
   before(function () {
     cy.visit('/')
-    cy.blockAllFeatureFlag()
     cy.exitWallet()
     cy.createEmptyWallet(true)
   })
 
   it('should check network detail without switching network', function () {
-    cy.getByTestID('portfolio_list').should('exist').wait(3000)
-    cy.getByTestID('header_active_network').first().invoke('text').then((network) => {
-      cy.getByTestID('header_status_indicator').should('have.css', 'background-color').then((statusBgColor) => {
-        cy.getByTestID('portfolio_header_container').filter(':visible').click()
-        cy.getByTestID('network_details_network').should('exist').contains(network)
-        cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', statusBgColor)
-      })
+    cy.getByTestID('bottom_tab_portfolio').click()
+    cy.getByTestID('header_settings').click()
+    cy.getByTestID('header_active_network').first().invoke('text').then((network: string) => {
+      cy.getByTestID('header_network_icon').filter(':visible').click()
+      cy.getByTestID('network_details_network').should('exist').contains(network)
+      cy.getByTestID(`button_network_${network}_check`).should('exist')
+      cy.getByTestID(`button_network_${network}_check`).should('have.css', 'color', 'rgb(0, 173, 29)')
     })
   })
 
   it('should check network detail by switching network', function () {
     cy.getByTestID('bottom_tab_portfolio').click()
     cy.getByTestID('header_settings').click()
-    cy.getByTestID('button_selected_network').should('exist').wait(3000)
-    cy.getByTestID('header_active_network').first().invoke('text').then((network) => {
-      cy.getByTestID('header_status_indicator').should('have.css', 'background-color').then((statusBgColor) => {
-        cy.getByTestID('setting_header_container').filter(':visible').click()
+    cy.getByTestID('header_active_network').first().invoke('text').then((network: string) => {
+      cy.getByTestID('header_network_icon').find('path').should('have.css', 'fill').then((statusBgColor) => {
+        cy.getByTestID('header_network_icon').filter(':visible').click()
         cy.getByTestID('network_details_network').should('exist').contains(network)
-        cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', statusBgColor)
-        cy.getByTestID('network_details_header_back').filter(':visible').click()
-        cy.getByTestID('button_selected_network').click()
+        cy.getByTestID(`button_network_${network}_check`).should('have.css', 'color', statusBgColor)
+        cy.go('back')
+        cy.getByTestID('header_network_icon').click()
         cy.getByTestID('button_network_Playground').click()
-        cy.blockAllFeatureFlag()
         cy.exitWallet()
         cy.createEmptyWallet(true)
         cy.getByTestID('header_settings').click()
-        cy.getByTestID('button_selected_network').should('exist').wait(3000)
-        cy.getByTestID('header_active_network').first().invoke('text').then((updatedNetwork) => {
-          cy.getByTestID('header_status_indicator').should('have.css', 'background-color').then((updatedStatusBgColor) => {
-            cy.getByTestID('setting_header_container').filter(':visible').click().wait(3000)
+        cy.getByTestID('header_network_icon').should('exist').wait(3000)
+        cy.getByTestID('header_active_network').first().invoke('text').then((updatedNetwork: string) => {
+          cy.getByTestID('header_network_icon').find('path').should('have.css', 'fill').then((updatedStatusBgColor) => {
+            cy.getByTestID('header_network_icon').click().wait(3000)
             expect(network).not.eq(updatedNetwork)
-            cy.getByTestID('network_details_status_value').should('exist').contains('Connected')
             cy.getByTestID('network_details_network').should('exist').contains(updatedNetwork)
-            cy.getByTestID('network_details_status_icon').should('have.css', 'background-color', updatedStatusBgColor)
+            cy.getByTestID(`button_network_${updatedNetwork}_check`).should('have.css', 'color', updatedStatusBgColor)
           })
         })
       })
@@ -172,14 +152,16 @@ context('Wallet - Network detail screen - with wallet context', () => {
 
   it('should be able to click block height and redirect it to defiscan', function () {
     cy.getByTestID('bottom_tab_portfolio').click()
-    cy.getByTestID('header_settings').click()
-    cy.getByTestID('setting_header_container').filter(':visible').click()
+    cy.getByTestID('header_settings').click().wait(3000)
+    cy.getByTestID('header_network_icon').should('exist').click()
+    cy.getByTestID('block_detail_explorer_url').invoke('text').then((block) => {
+      cy.getByTestID('block_detail_explorer_url').filter(':visible').click()
+    })
   })
 })
 
 context('Wallet - Network detail screen - with wallet context go back check', () => {
   before(function () {
-    cy.blockAllFeatureFlag()
     cy.visit('/')
     cy.exitWallet()
     cy.createEmptyWallet(true)
@@ -189,7 +171,7 @@ context('Wallet - Network detail screen - with wallet context go back check', ()
     cy.getByTestID('bottom_tab_portfolio').click().wait(3000)
     cy.url().should('include', 'app/portfolio')
     cy.getByTestID('portfolio_header_container').filter(':visible').click().wait(3000)
-    cy.getByTestID('network_details_header_back').click()
+    cy.go('back')
     cy.url().should('include', 'app/portfolio')
   })
 
@@ -198,7 +180,7 @@ context('Wallet - Network detail screen - with wallet context go back check', ()
     cy.url().should('include', 'app/DEX/DexScreen')
     cy.getByTestID('dex_guidelines_screen').should('exist')
     cy.getByTestID('dex_header_container').filter(':visible').click().wait(3000)
-    cy.getByTestID('network_details_header_back').click()
+    cy.go('back')
     cy.url().should('include', 'app/DEX/DexScreen')
   })
 
@@ -208,25 +190,26 @@ context('Wallet - Network detail screen - with wallet context go back check', ()
     cy.getByTestID('close_dex_guidelines').click().wait(3000)
     cy.url().should('include', 'app/DEX/DexScreen')
     cy.getByTestID('dex_header_container').filter(':visible').click().wait(3000)
-    cy.getByTestID('network_details_header_back').click()
+    cy.go('back')
     cy.url().should('include', 'app/DEX/DexScreen')
   })
 
   it('should get back to the transaction page when network detail called from transaction page', function () {
     cy.getByTestID('bottom_tab_portfolio').click()
     cy.getByTestID('switch_account_button').click()
-    cy.url().should('include', 'app/transactions')
-    cy.getByTestID('transactions_header_container').filter(':visible').click().wait(3000)
-    cy.getByTestID('network_details_header_back').click()
-    cy.url().should('include', 'app/transactions')
+    cy.getByTestID('transactions_tab').click()
+    cy.url().should('include', 'app/TransactionsScreen')
+    cy.getByTestID('portfolio_header_container').filter(':visible').click().wait(3000)
+    cy.go('back')
+    cy.url().should('include', 'app/TransactionsScreen')
   })
 
   it('should get back to the setting page when network detail called from setting page', function () {
     cy.getByTestID('bottom_tab_portfolio').click().wait(3000)
     cy.getByTestID('header_settings').filter(':visible').click().wait(3000)
     cy.url().should('include', 'app/Settings')
-    cy.getByTestID('setting_header_container').filter(':visible').click().wait(3000)
-    cy.getByTestID('network_details_header_back').click()
+    cy.getByTestID('header_network_icon').filter(':visible').click().wait(3000)
+    cy.go('back')
     cy.url().should('include', 'app/Settings')
   })
 })
