+++ mobile-app/cypress/e2e/functional/wallet/transaction/detail.spec.ts
@@ -4,7 +4,7 @@ context('Wallet - Transaction - Detail', () => {
     cy.sendDFItoWallet().wait(4000)
     cy.getByTestID('bottom_tab_portfolio').click()
     cy.getByTestID('switch_account_button').click()
-    cy.getByTestID('bottom_tab_transactions').click()
+    cy.getByTestID('transactions_tab').click()
     cy.getByTestID('transaction_row_0').click()
   })
