+++ mobile-app/cypress/e2e/functional/wallet/transaction/list_empty.spec.ts
@@ -4,7 +4,7 @@ context('Wallet - Transaction - Empty', () => {
       cy.createEmptyWallet(true)
       cy.getByTestID('bottom_tab_portfolio').click()
       cy.getByTestID('switch_account_button').click()
-      cy.getByTestID('bottom_tab_transactions').click()
+      cy.getByTestID('transactions_tab').click()
     })
 
     it('should display 0 row', () => {
